update more for ForgotPassword

+++ frontend/src/Services/utils.js
@@ -1,10 +1,13 @@
-export const registerUser  = (email, password) => {
+export const registerUser  = (email, password, additionalInfo) => {
     const storedUsers = JSON.parse(localStorage.getItem('users')) || [];
-    const newUser  = { email, password };
-    storedUsers.push(newUser );
-    localStorage.setItem('users', JSON.stringify(storedUsers));
+    
+    const newUser  = { email, password, ...additionalInfo }; 
+    const emailExists = storedUsers.some(user => user.email === email);
+    if (!emailExists) {
+        storedUsers.push(newUser );
+        localStorage.setItem('users', JSON.stringify(storedUsers));
+    }
 
-    // Lưu email vào sessionStorage
     const storedEmails = JSON.parse(localStorage.getItem('email')) || [];
     if (!storedEmails.includes(email)) {
         storedEmails.push(email);
